+++ CHANGES.rst
@@ -17,6 +17,7 @@ New features
 Enhancements
 ------------
 - Allow user to specify periodicity and box size per dimension [#276]
+- Allow larger Rmax (up to half the boxsize) [#277]
 
 Changes
 -------